requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
aiomysql>=0.2.0
pyyaml==6.0.1
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
